Create Switch Case-2
You are designing a basic menu-driven banking system for managing customer transactions. The menu allows the user to perform the following actions:
Check Balance
Deposit Money
Withdraw Money
Exit
Your task is to implement this menu and handle user selections using a switch-case structure.
Question:
Write a C program that displays the menu to the user, takes their choice, and performs the corresponding action. Assume an initial balance of $1000.